requirements.txt
streamlit
ddgs
python-dotenv
langchain
transformers
torch
faiss-cpu
sentence-transformers
duckduckgo-search
langchain-perplexity
